Struct x11_get_windows::Session

pub struct Session {
    pub display: Display,
    pub root_window: Option<Window>,
    pub client_list_atom: Option<Atom>,
    pub active_window_atom: Option<Atom>,
}

This is meant to be a struct that makes it easy to use this crate.

This is purely for convenience.

Example

let mut session = Session::open()
   .expect("Error opening a new session.");
session
   .get_windows()
   .expect("Could not get a list of windows.")
   .iter()
   .filter_map(|x| x.get_title(&session.display).ok())
   .for_each(|x| println!("{:?}", x.as_ref()))
// Prints out the title for every window that is visible on the screen.

Fields
display: Display

A display that has been opened.

root_window: Option<Window>

The root window of the display.

client_list_atom: Option<Atom>

The atom that represents the client_list property.

active_window_atom: Option<Atom>

The atom that represents the active_window property.

Methods

impl Session

pub fn open() -> Result<Self, Null>

Opens a display.

pub fn from_display(display: Display) -> Self

Creates a session from an already opened Display connection.

See Display::open for more information.

pub fn get_windows(&mut self) -> Result<Windows, NotSupported>

Gets all the current windows on the screen.

This will update any values that are set to None if it needs to use them.

This can possible produce a NotSupported error. In that case, please read the documentation for that struct.

pub fn active_window(&mut self) -> Result<Window, NotSupported>

Gets the currently active window in the display.